requirements.txt
aiohttp==3.7.3
aiojobs==0.3.0
alembic==1.6.5
astroid==2.4.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
cleanco==2.1
click==7.1.2
databases==0.4.1
dill==0.3.4
docopt==0.6.2
# fastapi==0.63.0
fastapi==0.68.1
fastapi-plugins==0.8.1
h11==0.12.0
httpcore<0.14.0
httpx==0.19.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.0
pandarallel==1.5.1
pandas==1.2.5
pangres==2.2.3
pipreqs==0.4.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydantic==1.8.2
PyJWT==2.1.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
rapidfuzz==1.4.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.23
starlette==0.14.2
tenacity==7.0.0
toml==0.10.2
typing-extensions==3.7.4.3
ujson==3.1.0
urllib3==1.26.3
# uvicorn==0.13.3
wrapt==1.12.1
yarg==0.1.9
yarl==1.6.3
scikit-learn==0.23.2
statsmodels==0.12.0
wbgapi==1.0.5
gunicorn==20.1.0
uvicorn==0.15.0
google-cloud-logging==2.6.0
msgpack==1.0.2
aioredis==2.0.0
# uvloop==0.16.0
# tomli==1.2.1
py15rock==0.0.24
